refactor: Rename `HorizontalWeightPicker` to `HorizontalSlider`
This commit renames the `horizontal_weight_picker.dart` file to `horizontal_slider.dart` to better reflect its general-purpose functionality as a horizontal slider rather than a specific weight picker. This change aligns with the widget's ability to select a value within a range, as opposed to being strictly limited to weight selection.